requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-decouple==2.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
twilio==7.9.3
urllib3==1.26.9
whitenoise==6.2.0
